backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.3
celery==5.3.4
python-dotenv==1.0.1
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.24
pymilvus==2.4.4
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.3
requests-html==0.10.0
playwright==1.40.0
openai==1.12.0
anthropic==0.18.0
zhipuai==2.0.1
dashscope==1.13.4
alipay-sdk-python==3.14.0
wechatpay-apiv3==0.3.4
aiofiles==23.2.1
aiosmtplib==3.0.1
jinja2==3.1.3
httpx==0.27.0
tenacity==8.2.3
 structlog==24.1.0
sentry-sdk[fastapi]==1.40.6
prometheus-client==0.19.0
